Rename backend list modules for consistency
